Use stroke-mixin for geometric data values
